app-layer: user shorter integer for array access

+++ src/app-layer-ssl.c
@@ -1393,7 +1393,7 @@ static uint32_t GetCertsLen(SSLStateConnp *curr_connp, const uint8_t *input,
             return len;
         } else if (input_len + curr_connp->trec_pos >= 3) {
             uint8_t buf[3] = { 0, 0, 0, }; // init to 0 to make scan-build happy
-            uint32_t i = 0;
+            uint8_t i = 0;
             for (uint32_t x = 0; x < curr_connp->trec_pos && i < 3;  x++) {
                 buf[i++] = curr_connp->trec[x];
             }
